Rename test files to test_ prefix
