+++ tests/posit/connect/test_auth.py
