+++ tests/posit/connect/test_client.py
