+++ tests/posit/connect/test_errors.py
